ocr_text/requirements.txt
pillow
filetype==1.2.0
paddlepaddle==3.0.0
paddleocr==3.0.1
